Remove old style ecl tests
