+++ src/ert/shared/share/ert/forward-models/res/script/ecl300_config.yml
@@ -1,30 +1,5 @@
 # This is an example configuration file for the local ecl300 installation, see
-# the ecl100_config.yml file for more extensive dcoujmentation.
-
-versions:
-  '2015.2':  # without quotes this will be interpreted as a number
-    scalar:
-      executable: /path/ecl/2015.2/linux/x86_64/bin/eclipse300
-    mpi:
-      executable: /path/ecl/2015.2/linux/x86_64/bin/eclipse300_mpi
-      mpirun: /path/ecl/2015.2/intel/bin/mpirun
-      env:
-        MPI_ROOT: /path/to/mpi/root
-        LD_LIBRARY_PATH: /path/to/mpi/root/lib64:$LD_LIBRARY_PATH
-
-  '2017.2':
-    scalar:
-      executable: /path/ecl/2017.2/linux/x86_64/bin/eclipse300
-      env:
-    mpi:
-      executable: /path/ecl/2017.2/linux/x86_64/bin/eclipse300_mpi
-      mpirun: /path/ecl/2017.2/intel/bin/mpirun
-
-
-env:
-  LM_LICENSE_FILE: license@company.com
-  F_UFMTENDIAN: big
-  ARCH: x86_64
+# the ecl100_config.yml file for more extensive documentation.
 
 eclrun_env:
   SLBSLS_LICENSE_FILE: something@yourcompany.com
